requirements.txt
Flask==2.3.3
Flask-Cors==6.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
Werkzeug==3.1.3
psycopg==3.3.2
python-dotenv==1.1.1
langchain==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.0.1
google-generativeai==0.8.3
langchain-aws==0.2.9
pandas==2.2.3
numpy==1.26.4
crewai==1.8.0
crewai-tools==0.76.0
pypdf2==3.0.1
ddgs==9.10.0